refactor mint page
